Make it possible to spend hp to move through monsters
Previously it was only possible to spend swords, now it's possible to spend health (if available) or swords to move.

Cleaned up a few comments and marked important ones as TODO.

Refactored ActionGenerator out of RandomPlanner for later re-use.  Also makes it possible for ActionGenerator to eventually be "trusted".

Generate all possible card plays up front, even though we only pick the first one.  This will be important once we have CardPlays which have multiple-choice side effects and we'll want to randomly pick from such.