Add CardSubType and support Gems
This is in preparation for  the Gem Collector companion which (is the one card which) cares if other Cards are Gems or not.

Added new bin/box_contents.dart to spit out static numbers about the cards (and probably eventually loot) as I add them).

Also fixed formatting in README a little.