Add triggered effects "if have X do Y"
These trigger any time during the turn, so the cards just add the effects to a queue and the queue is checked after every action.  Once triggered effects are removed from the queue.

This made possible all Rebel companions, Kobold Merchant, Queen of Hearts, King of the Mountain, Wand of Recall and Archaeologist.

Each card was (separately) unit tested as they're all a little complex.