Fix some secrets to discard immediately
Previously I implemented secrets as either being held to the end of the game, or held until use.  However in more a more careful reading of the rules (and boardgamegeek) it's clear that some secrets take effect immediately (e.g. acquire effects) and then are immediately discarded.  Implemented this as ItemMode.

Added a bunch of asserts around usage of items to prevent "keep" or "discardImmediately" items from every being used.

Added unit tests for all the new discardImmediately behavior.