Add Swagger and Search
These are the last two cards!  Now correctly implementing 183 of 183 cards!

Swagger's implementation is currently sensitive to card ordering.  Could implement it differently with a tiny amount of work, but this was easier for now.

Wrote the pseudo-code for placing monkey idols, but commented it out for now since without the "deep" half of the map added, there is no where to put them and the code would crash.

Added unit test for Swagger and Search.