Fix trash exceptions
Fix exceptions seen during TrashCard.  It turns out that trashCard shouldn't ever really fail with an exception as currently designed.  Master Burglar is find to use when you don't have a burgle, and Magic Spring sometimes can fail if you've reshuffled a card out of your discard before the trash takes effect at the end of turn (unclear what the rule here should be?)

Fixed simulate.dart handling of seed argument to apply to planners as well and thus actually work.

Add a toString to all actions and use a single print statement to cover all actions (easier to comment out or switch away from print in the future).

Also discovered that location checks were being made at the wrong time in the execute side.  Dragon's Eye is valid to play even outside the depths for example, but only is valid to acquire inside the depths.  Fixed and tested.